requirements.txt
aiofiles==24.1.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
blinker==1.9.0
certifi==2025.4.26
click==8.2.0
colorama==0.4.6
configparser==7.2.0
dnspython==2.7.0
email_validator==2.2.0
Faker==37.3.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
python-jose[cryptography]
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
